update bkg split into train and test
