+++ environment.yaml
@@ -23,3 +23,4 @@ dependencies:
     - wandb==0.17.3
     - scipy==1.14.1
     - einops==0.8.0
+    - nflows-0.14
