+++ gwak/train/cl_models.py
@@ -20,8 +20,8 @@ from typing import Optional
 import torch
 import torch.nn as nn
 from einops import rearrange, repeat
-from losses import SupervisedSimCLRLoss
-from schedulers import WarmupCosineAnnealingLR
+from gwak.train.losses import SupervisedSimCLRLoss
+from gwak.train.schedulers import WarmupCosineAnnealingLR
 
 import numpy as np
 import matplotlib.pyplot as plt
@@ -62,251 +62,6 @@ class ModelCheckpoint(pl.callbacks.ModelCheckpoint):
 
         with open(os.path.join(save_dir, "model_JIT.pt"), "wb") as f:
             torch.jit.save(trace, f)
-
-
-class LinearModelCheckpoint(pl.callbacks.ModelCheckpoint):
-
-    def on_train_end(self, trainer, pl_module):
-        torch.cuda.empty_cache()
-
-        module = pl_module.__class__.load_from_checkpoint(
-            self.best_model_path,
-            **pl_module.hparams['init_args']
-        )
-
-        # Modifiy these to read the last training/valdation data
-        # to acess the input shape.
-        X = torch.randn(1, 8).to(module.device) # GWAK 2
-        trace = torch.jit.trace(module.model.to("cpu"), X.to("cpu"))
-
-        save_dir = trainer.logger.log_dir or trainer.logger.save_dir
-
-        with open(os.path.join(save_dir, "mlp_model_JIT.pt"), "wb") as f:
-            torch.jit.save(trace, f)
-
-
-class Linear(GwakBaseModelClass):
-    def __init__(
-            self,
-            backgrounds: str = '/home/hongyin.chen/whiten_timeslide.h5', # timeslides to train against
-            ckpt: str = "../output/test_S4_fixedSignals_0p5sec_v2/lightning_logs/2wla29uz/checkpoints/27-1400.ckpt",
-            cfg_path: str = "../output/test_S4_fixedSignals_0p5sec_v2/config.yaml", # pre-trained embedding model
-            new_shape=128,
-            n_dims=16,
-            learning_rate=1e-3):
-        super().__init__()
-        self.model = nn.Linear(n_dims, 1)
-
-        with h5py.File(backgrounds, "r") as h:
-            # Load data as a NumPy array first
-            data = h["data"][:128*10*20, :, :1024]  # Ensuring it's a NumPy array
-
-        # Splitting into train and validation sets (80% train, 20% validation)
-        self.backgrounds, self.val_backgrounds = train_test_split(data, test_size=0.2, random_state=42)
-        self.backgrounds = torch.from_numpy(self.backgrounds).to("cpu")
-        self.val_backgrounds = torch.from_numpy(self.val_backgrounds).to("cpu")
-
-        self.ckpt = ckpt
-        self.cfg_path = cfg_path
-        # Load first model (frozen for inference)
-        with open(self.cfg_path,"r") as fin:
-            cfg = yaml.load(fin,yaml.FullLoader)
-        self.graph = Crayon.load_from_checkpoint(self.ckpt, **cfg['model']['init_args'])
-        self.graph.eval()
-        self.graph.to(device=self.device)
-
-        self.new_shape = new_shape
-        self.learning_rate = learning_rate
-
-        self.save_hyperparameters()
-
-    def forward(self, x):
-        return self.model(x)
-
-    def configure_optimizers(self):
-        return torch.optim.Adam(self.model.parameters(), lr=self.learning_rate)
-
-    def training_step(self, batch, batch_idx):
-        i = batch_idx  # Each batch corresponds to a background chunk
-        small_bkgs = self.backgrounds[i * self.new_shape:(i + 1) * self.new_shape].to(self.device)
-
-        # Load first model (frozen for inference)
-        batch = self.graph.model(batch[0])
-        small_bkgs = self.graph.model(small_bkgs)
-
-        background_MV = self.model(small_bkgs)
-        signal_MV = self.model(batch)
-
-        # Take min score for each signal example
-        signal_MV = torch.min(signal_MV, dim=1)[0]
-
-        zero = torch.tensor(0.0, device=self.device)
-
-        background_loss = torch.maximum(zero, 1 - background_MV).mean()
-        signal_loss = torch.maximum(zero, 1 + signal_MV).mean()
-
-        loss = background_loss + signal_loss
-
-        self.log(
-            'train_loss',
-            loss,
-            on_epoch=True, # Newly added
-            sync_dist=True)
-
-        return loss
-
-    @torch.no_grad()
-    def validation_step(self, batch, batch_idx):
-        """Validation step processes all validation data at once."""
-        small_bkgs = self.val_backgrounds.to(self.device)
-
-        # Load first model (frozen for inference)
-        batch = self.graph.model(batch[0])
-        small_bkgs = self.graph.model(small_bkgs)
-
-        background_MV = self.model(small_bkgs)
-        signal_MV = self.model(batch)
-
-        # Take min score for each signal example
-        signal_MV = torch.min(signal_MV, dim=1)[0]
-
-        zero = torch.tensor(0.0, device=self.device)
-
-        background_loss = torch.maximum(zero, 1 - background_MV).mean()
-        signal_loss = torch.maximum(zero, 1 + signal_MV).mean()
-
-        loss = background_loss + signal_loss
-        self.log(
-            'val/loss',
-            loss,
-            sync_dist=True)
-
-        return loss
-
-    def configure_callbacks(self) -> Sequence[pl.Callback]:
-        # checkpoint for saving best model
-        # that will be used for downstream export
-        # and inference tasks
-        # checkpoint for saving multiple best models
-        callbacks = []
-
-        # if using ray tune don't append lightning
-        # model checkpoint since we'll be using ray's
-        checkpoint = LinearModelCheckpoint(
-            monitor='val/loss',
-            save_last=True,
-            auto_insert_metric_name=False
-            )
-
-        callbacks.append(checkpoint)
-
-        return callbacks
-
-class MLPModel(nn.Module):
-    def __init__(self, input_dim, hidden_dim=32):
-        super().__init__()
-        self.fc1 = nn.Linear(input_dim, hidden_dim)  # Hidden layer
-        self.fc2 = nn.Linear(hidden_dim, 1)  # Output layer
-        self.activation = nn.ReLU()  # Non-linearity
-        self.sigmaboy = nn.Sigmoid()
-
-    def forward(self, x):
-        x = self.activation(self.fc1(x))  # Apply activation
-        x = self.sigmaboy(self.fc2(x))  # Output score (no sigmoid)
-        return x
-
-class NonLinearClassifier(GwakBaseModelClass):
-    def __init__(
-            self,
-            backgrounds: str = '/home/hongyin.chen/whiten_timeslide.h5',
-            ckpt: str = "../output/test_S4_fixedSignals_0p5sec_v2/lightning_logs/2wla29uz/checkpoints/27-1400.ckpt",
-            cfg_path: str = "../output/test_S4_fixedSignals_0p5sec_v2/config.yaml",
-            new_shape=128,
-            n_dims=16,
-            learning_rate=1e-3,
-            hidden_dim=32):
-        super().__init__()
-        self.model = MLPModel(n_dims, hidden_dim)
-
-        with h5py.File(backgrounds, "r") as h:
-            data = h["data"][:128*10*20, :, :1024]
-
-        self.backgrounds, self.val_backgrounds = train_test_split(data, test_size=0.2, random_state=42)
-        self.backgrounds = torch.tensor(self.backgrounds, dtype=torch.float32).to("cpu")
-        self.val_backgrounds = torch.tensor(self.val_backgrounds, dtype=torch.float32).to("cpu")
-
-        self.ckpt = ckpt
-        self.cfg_path = cfg_path
-
-        with open(self.cfg_path, "r") as fin:
-            cfg = yaml.load(fin, yaml.FullLoader)
-        self.graph = Crayon.load_from_checkpoint(self.ckpt, **cfg['model']['init_args'])
-        self.graph.eval()
-        self.graph.to(device=self.device)
-
-        self.new_shape = new_shape
-        self.learning_rate = learning_rate
-
-        self.save_hyperparameters()
-
-    def forward(self, x):
-        return self.model(x)
-
-    def configure_optimizers(self):
-        return optim.Adam(self.model.parameters(), lr=self.learning_rate)
-
-    def training_step(self, batch, batch_idx):
-        i = batch_idx
-        small_bkgs = self.backgrounds[i * self.new_shape:(i + 1) * self.new_shape].to(self.device)
-
-        # Feature extraction
-        signal_feats = self.graph.model(batch[0])
-        background_feats = self.graph.model(small_bkgs)
-
-        # Predictions
-        signal_preds = self.forward(signal_feats)
-        background_preds = self.forward(background_feats)
-
-        # BCE targets: 1 = signal, 0 = background
-        target_signal = torch.ones_like(signal_preds)
-        target_background = torch.zeros_like(background_preds)
-
-        # Compute BCE loss
-        loss_signal = F.binary_cross_entropy(signal_preds, target_signal)
-        loss_background = F.binary_cross_entropy(background_preds, target_background)
-        loss = loss_signal + loss_background
-
-        self.log("train_loss", loss, on_epoch=True, sync_dist=True)
-
-        return loss
-
-    @torch.no_grad()
-    def validation_step(self, batch, batch_idx):
-        small_bkgs = self.val_backgrounds.to(self.device)
-
-        signal_feats = self.graph.model(batch[0])
-        background_feats = self.graph.model(small_bkgs)
-
-        signal_preds = self.forward(signal_feats)
-        background_preds = self.forward(background_feats)
-
-        target_signal = torch.ones_like(signal_preds)
-        target_background = torch.zeros_like(background_preds)
-
-        loss_signal = F.binary_cross_entropy(signal_preds, target_signal)
-        loss_background = F.binary_cross_entropy(background_preds, target_background)
-        loss = loss_signal + loss_background
-
-        self.log("val/loss", loss, sync_dist=True)
-
-        return loss
-
-    def configure_callbacks(self) -> Sequence[pl.Callback]:
-        return [LinearModelCheckpoint(
-            monitor='val/loss',
-            save_last=True,
-            auto_insert_metric_name=False
-        )]
 
 class Encoder(nn.Module):
 
@@ -603,26 +358,6 @@ class Autoencoder(GwakBaseModelClass):
         callbacks.append(checkpoint)
 
         return callbacks
-
-
-class gwak2(GwakBaseModelClass):
-
-    def __init__(self, ):
-
-        super().__init__()
-
-    def training_step(self, batch, batch_idx):
-        self.log(
-            'train_loss',
-            loss,
-            sync_dist=True)
-
-        return loss
-
-    def configure_optimizers(self):
-        optimizer = optim.Adam(self.parameters(), lr=config.learning_rate)
-        return optimizer
-
 
 class DropoutNd(nn.Module):
     def __init__(self, p: float = 0.5, tie=True, transposed=True):
