+++ gwak/train/configs/FM_multiSignalAndBkg.yaml
@@ -15,7 +15,7 @@ trainer:
   accumulate_grad_batches: 1
 ckpt_path: null
 model:
-  class_path: models.NonLinearClassifier
+  class_path: fm_models.NonLinearClassifier
   init_args:
     backgrounds: /home/hongyin.chen/whiten_timeslide.h5
     ckpt: /home/katya.govorkova/gwak2/gwak/output/test_S4_fixedSignals_0p5sec_v2/lightning_logs/2wla29uz/checkpoints/33-1700.ckpt
@@ -72,7 +72,7 @@ data:
       - null
       - null
       - null
-    data_dir: /home/katya.govorkova/gwak2_background
+    data_dir: /home/katya.govorkova/gwak2_background/train
     sample_rate: 2048
     kernel_length: 0.5
     psd_length: 64
