+++ gwak/train/configs/NF_onlyBkg.yaml
@@ -1,0 +1,32 @@
+# lightning.pytorch==2.3.2
+seed_everything: 0
+trainer:
+  callbacks:
+    - class_path: callback.ValidationCallback
+  accelerator: gpu
+  devices: [1]
+  # accelerator: auto
+  # devices: auto
+  strategy: ddp_find_unused_parameters_true
+  logger:
+    class_path: lightning.pytorch.loggers.WandbLogger
+  max_epochs: 50
+  check_val_every_n_epoch: 1
+  accumulate_grad_batches: 1
+ckpt_path: null
+model:
+  class_path: fm_models.BackgroundFlowModel
+
+data:
+  class_path: dataloader.TimeSlidesDataloader
+  init_args:
+    data_dir: /home/katya.govorkova/gwak2_background/train
+    sample_rate: 2048
+    kernel_length: 0.5
+    psd_length: 64
+    fduration: 1
+    fftlength: 2
+    batch_size: 128
+    batches_per_epoch: 10
+    num_workers: 2
+    data_saving_file: null
