+++ gwak/train/configs/Transformer_SimCLR_multiSignalAndBkg_noSG.yaml
@@ -16,7 +16,7 @@ trainer:
   accumulate_grad_batches: 1
 ckpt_path: null
 model:
-  class_path: models.Tarantula
+  class_path: cl_models.Tarantula
   init_args:
     num_ifos: 2
     num_timesteps: 2048
@@ -82,7 +82,7 @@ data:
       - null
       - null
       - null
-    data_dir: /n/holystore01/LABS/iaifi_lab/Lab/sambt/LIGO/gwak2_background
+    data_dir: /n/netscratch/iaifi_lab/Lab/emoreno/O4_MDC_background/train
     sample_rate: 4096
     kernel_length: 0.5
     psd_length: 64
