+++ gwak/train/configs/Transformer_SimCLR_multiSignal_all.yaml
@@ -20,7 +20,7 @@ optimizer:
   init_args:
     lr: 1e-4
 model:
-  class_path: models.Tarantula
+  class_path: cl_models.Tarantula
   init_args:
     num_ifos: 2
     num_timesteps: 2048
@@ -85,7 +85,7 @@ data:
       - null
       - null
       - null
-    data_dir: /n/holystore01/LABS/iaifi_lab/Lab/sambt/LIGO/gwak2_background
+    data_dir: /n/netscratch/iaifi_lab/Lab/emoreno/O4_MDC_background/train
     sample_rate: 4096
     kernel_length: 0.5
     psd_length: 64
